Combat Mechanics
More or less finished the CombatManager update loop.
Merged together the Move and Alteration systems under ActorBehaviour.
Turned into Monobehaviour a number of previously default c# class to enable use of GetComponentInParent and such utilities.
Started to create a first effect and targeting so that testing the combat becomes possible.