Combat Start
Tried to deal with the start of combat but did shit. Need to solve the events shit for the UI, and to decide what sould deal with the set-up for the start of combat (CombatManager or a new Class? In the Start or in a dedicated methode ? etc), bref need more thought before anything can be done